backend/requirements.txt
bcrypt==3.2.0
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-JWT-Extended==3.25.0
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.15.0
six==1.15.0
SQLAlchemy==1.3.22
Werkzeug==1.0.1
